Trochę mniej exceptions przy próbie ewaluacji

+++ przypominajka/src/remembrall/nodes/ConstrNode.java
@@ -31,28 +31,28 @@ public class ConstrNode implements Node {
 //			datetime(int day, int month, int year, int hour, int min)
 			if (args.size() == 3)
 				return new IdentValue(new Datetime(
-						(Integer)args.get(0).evalNode(env).v,
-						(Integer)args.get(1).evalNode(env).v,
-						(Integer)args.get(2).evalNode(env).v));
+						(Long)args.get(0).evalNode(env).v,
+						(Long)args.get(1).evalNode(env).v,
+						(Long)args.get(2).evalNode(env).v));
 			if (args.size() == 4 && args.get(3) instanceof Time)
 				return new IdentValue(new Datetime(
-						(Integer)args.get(0).evalNode(env).v,
-						(Integer)args.get(1).evalNode(env).v,
-						(Integer)args.get(2).evalNode(env).v,
+						(Long)args.get(0).evalNode(env).v,
+						(Long)args.get(1).evalNode(env).v,
+						(Long)args.get(2).evalNode(env).v,
 						(Time)args.get(3).evalNode(env).v));
 			if (args.size() == 4)
 				return new IdentValue(new Datetime(
-						(Integer)args.get(0).evalNode(env).v, 
-						(Integer)args.get(1).evalNode(env).v, 
-						(Integer)args.get(2).evalNode(env).v,
-						(Integer)args.get(3).evalNode(env).v));
+						(Long)args.get(0).evalNode(env).v, 
+						(Long)args.get(1).evalNode(env).v, 
+						(Long)args.get(2).evalNode(env).v,
+						(Long)args.get(3).evalNode(env).v));
 			if (args.size() == 5)
 				return new IdentValue(new Datetime(
-						(Integer)args.get(0).evalNode(env).v, 
-						(Integer)args.get(1).evalNode(env).v, 
-						(Integer)args.get(2).evalNode(env).v,
-						(Integer)args.get(3).evalNode(env).v,
-						(Integer)args.get(4).evalNode(env).v));
+						(Long)args.get(0).evalNode(env).v, 
+						(Long)args.get(1).evalNode(env).v, 
+						(Long)args.get(2).evalNode(env).v,
+						(Long)args.get(3).evalNode(env).v,
+						(Long)args.get(4).evalNode(env).v));
 		case typeLocation:
 			return new IdentValue(new Location(
 					(String)args.get(0).evalNode(env).v, 
@@ -63,15 +63,15 @@ public class ConstrNode implements Node {
 		case typeTime:
 			if (args.size() == 1)
 				return new IdentValue(new Time(
-						(Integer)args.get(0).evalNode(env).v));
+						(Long)args.get(0).evalNode(env).v));
 			if (args.size() == 2)
 				return new IdentValue(new Time(
-						(Integer)args.get(0).evalNode(env).v, 
-						(Integer)args.get(1).evalNode(env).v));
+						(Long)args.get(0).evalNode(env).v, 
+						(Long)args.get(1).evalNode(env).v));
 			return new IdentValue(new Time(
-					(Integer)args.get(0).evalNode(env).v, 
-					(Integer)args.get(1).evalNode(env).v, 
-					(Integer)args.get(2).evalNode(env).v));
+					(Long)args.get(0).evalNode(env).v, 
+					(Long)args.get(1).evalNode(env).v, 
+					(Long)args.get(2).evalNode(env).v));
 		default:
 			return null;
 		}
